+++ rv/__init__.py
@@ -2,12 +2,4 @@
 
 __version__ = '0.4.0'
 
-ENCODING = 'cp1251'
-"""
-Encoding used to convert 8-bit strings to/from Unicode strings.
-
-SunVox uses the ``cp1251`` encoding which supports both US ASCII
-and Cyrillic scripts.
-
-See also `<https://en.wikipedia.org/wiki/Windows-1251>`__.
-"""
+ENCODING = 'utf8'